Migrate stabilized `Morph` functionality to `boilercv`. Fix docs nitpicks.
