+++ pipeline/boilercv_pipeline/correlations/dimensionless_bubble_diameter/types.py
@@ -14,6 +14,9 @@ from typing import (
 from numpy import float64
 from numpy.typing import NDArray
 from pydantic import PlainSerializer, PlainValidator
+from sympy import Basic, symbols, sympify
+
+from boilercv.morphs import Morph
 
 T = TypeVar("T")
 K = TypeVar("K")
@@ -45,6 +48,8 @@ Sym: TypeAlias = Literal["Fo_0", "Ja", "Re_b0", "Pr", "beta", "pi"]
 """Symbol."""
 syms: tuple[Sym, ...] = get_args(Sym)
 """Symbols."""
+solve_syms: tuple[Sym, ...] = ("Fo_0", "beta")
+"""Symbols to solve for."""
 FormsRepl: TypeAlias = Repl[Kind]
 """Forms replacements."""
 Param: TypeAlias = Literal[
@@ -71,3 +76,33 @@ Leaf: TypeAlias = Any
 """Leaf node."""
 Node: TypeAlias = dict[Any, "Node | Leaf"]
 """General node."""
+
+
+def validate_expr(v: Basic | str):
+    """Validate expression."""
+    return (
+        v
+        if isinstance(v, Basic)
+        else sympify(v, locals=LOCALS.model_dump(), evaluate=False)
+    )
+
+
+Expr: TypeAlias = Annotated[
+    Basic, PlainValidator(validate_expr), JsonStrPlainSerializer
+]
+"""Expression."""
+
+Locals: TypeAlias = Morph[Sym, Expr]
+"""Locals."""
+PARAMS = Morph[Param, Sym](dict(zip(params, syms, strict=True)))
+"""Parameters."""
+LOCALS = Locals(
+    dict(
+        zip(
+            syms,
+            symbols(list(PARAMS.values()), nonnegative=True, real=True, finite=True),
+            strict=False,
+        )
+    )
+)
+"""Local variables."""
