+++ src/boilercv/morphs/types.py
@@ -1,0 +1,67 @@
+"""Morph types."""
+
+from collections.abc import MutableMapping
+from typing import NamedTuple, ParamSpec, Protocol, TypeVar
+
+T = TypeVar("T", contravariant=True)
+R = TypeVar("R", covariant=True)
+P = ParamSpec("P")
+
+
+class TypeType(Protocol[T, R, P]):  # noqa: D101
+    def __call__(self, i: T, /, *args: P.args, **kwds: P.kwargs) -> R: ...  # noqa: D102
+
+
+K = TypeVar("K")
+V = TypeVar("V")
+
+
+class TypeMap(Protocol[T, K, V, P]):  # noqa: D101
+    def __call__(  # noqa: D102
+        self, i: T, /, *args: P.args, **kwds: P.kwargs
+    ) -> MutableMapping[K, V]: ...
+
+
+class TypeDict(Protocol[T, K, V, P]):  # noqa: D101
+    def __call__(  # noqa: D102
+        self, i: T, /, *args: P.args, **kwds: P.kwargs
+    ) -> dict[K, V]: ...
+
+
+class MapType(Protocol[K, V, R, P]):  # noqa: D101
+    def __call__(  # noqa: D102
+        self, i: MutableMapping[K, V], /, *args: P.args, **kwds: P.kwargs
+    ) -> R: ...
+
+
+class DictType(Protocol[K, V, R, P]):  # noqa: D101
+    def __call__(  # noqa: D102
+        self, i: dict[K, V], /, *args: P.args, **kwds: P.kwargs
+    ) -> R: ...
+
+
+RK = TypeVar("RK")
+RV = TypeVar("RV")
+
+
+class MapMap(Protocol[K, V, RK, RV, P]):  # noqa: D101
+    def __call__(  # noqa: D102
+        self, i: MutableMapping[K, V], /, *args: P.args, **kwds: P.kwargs
+    ) -> MutableMapping[RK, RV]: ...
+
+
+class DictDict(Protocol[K, V, RK, RV, P]):  # noqa: D101
+    def __call__(  # noqa: D102
+        self, i: dict[K, V], /, *args: P.args, **kwds: P.kwargs
+    ) -> dict[RK, RV]: ...
+
+
+KT = TypeVar("KT", bound=type)
+VT = TypeVar("VT", bound=type)
+
+
+class Types(NamedTuple):
+    """Inner types for mapings."""
+
+    key: type
+    value: type
